requirements.txt
setuptools
plumbum
PyYAML
